[FEAT/#56] 리팩토링 - 페이지를 위젯의 조합으로 분해
