+++ src/widgets/alarm-center/index.ts
@@ -1,0 +1,2 @@
+export { AlarmCenterHeader } from "./ui/AlarmCenterHeader";
+export { AlarmCenterSection } from "./ui/AlarmCenterSection";
